requirements.txt
playwright>=1.40.0
pydantic>=2.0.0
httpx>=0.24.0
pyyaml>=6.0
pytest>=8.0
anthropic>=0.40
